# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Market {
  id: number;
  cycle: number;
  date: string;
  venue: string;
  spend: number;
  revenue: number;
  margin: number;
}

export interface BuyItem {
  id: number;
  marketCycle: number;
  flower: string;
  detail: string;
  qty: number;
  unit: string;
  lastPrice: number;
  checked: boolean;
  category: string;
}

export interface BuyItemUpdate {
  checked: boolean;
}

export interface BouquetPlan {
  marketCycle: number;
  selectedBand: string;
  count: number;
}

export interface BouquetPlanUpdate {
  selectedBand: string;
  /** @minimum 0 */
  count: number;
}

export type CloseMarketCounts = {[key: string]: number};

export interface CloseMarket {
  marketCycle: number;
  counts: CloseMarketCounts;
  closed: boolean;
}

export type CloseMarketUpdateCounts = {[key: string]: number};

export interface CloseMarketUpdate {
  counts: CloseMarketUpdateCounts;
  closed: boolean;
}

export interface MarketContext {
  cycle: number;
  date: string;
  venue: string;
  spend: number;
  revenue: number;
  margin: number;
  buyItems: BuyItem[];
  bouquetPlan: BouquetPlan;
  closeMarket: CloseMarket;
}