# lib/api-zod/src/generated/types/bouquetPlanUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BouquetPlanUpdate {
  selectedBand: string;
  /** @minimum 0 */
  count: number;
}