# lib/api-zod/src/generated/types/buyItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BuyItem {
  id: number;
  marketCycle: number;
  flower: string;
  detail: string;
  qty: number;
  unit: string;
  lastPrice: number;
  checked: boolean;
  category: string;
}