# lib/api-zod/src/generated/types/closeMarket.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CloseMarketCounts } from './closeMarketCounts';

export interface CloseMarket {
  marketCycle: number;
  counts: CloseMarketCounts;
  closed: boolean;
}