# lib/api-zod/src/generated/types/closeMarketUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CloseMarketUpdateCounts } from './closeMarketUpdateCounts';

export interface CloseMarketUpdate {
  counts: CloseMarketUpdateCounts;
  closed: boolean;
}