# lib/api-zod/src/generated/types/market.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Market {
  id: number;
  cycle: number;
  date: string;
  venue: string;
  spend: number;
  revenue: number;
  margin: number;
}