# lib/api-zod/src/generated/types/marketContext.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BouquetPlan } from './bouquetPlan';
import type { BuyItem } from './buyItem';
import type { CloseMarket } from './closeMarket';

export interface MarketContext {
  cycle: number;
  date: string;
  venue: string;
  spend: number;
  revenue: number;
  margin: number;
  buyItems: BuyItem[];
  bouquetPlan: BouquetPlan;
  closeMarket: CloseMarket;
}